* updated squishy libs, that fixes a few bugs in IRC Interface concerning Nick cycling on connect * converted application into a Windows Application instead of Console Application to work around some limitations * console is created now by calling WINAPI, to allow custom font install + added a more readable custom font for console, it also looks kind of special comparing to the old boring Windows console and allows more text to fit the console window
It seems that using custom fonts in Windows console is poorly documented. I am using a Raster font, not a TrueType font for the console window.
Never saw an article or tutorial anywhere about this subject, except some hints how to add a new TrueType font, and nobody discovered how to change the font programmaticly without changing the properties of the console window yourself. My method should be handy for a variety of console based projects, it only involves setting some registry keys before the console is attached and calling a few WINAPI functions after it is attached.